+++ RCNutilidades/views/compras.hbs
@@ -32,7 +32,8 @@
     
             <div class="col-6" id="btn-add"> 
                 <button type="button" class="btn btn-primary">+ Novo</button>
-                <button type="button" class="btn btn-danger"><img src="/images/lata-lixo.png" alt="lata-lixo" style="width: 20px; padding: 1px;"></button>
+                <button type="button" class="btn btn-info" style="margin-left: 2px;"><i class="bi bi-pencil-fill"></i></button>
+                <button type="button" class="btn btn-danger" style="margin-left: 2px;"><img src="/images/lata-lixo.png" alt="lata-lixo" style="width: 20px; padding: 1px;"></button>
             </div>
         </div>
     </div>
@@ -44,50 +45,30 @@
                         <thead>
                         <tr class="header-prods-compras">
                             <th scope="col"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <th scope="col">NOME</th>
-                            <th scope="col">MARCA</th>
-                            <th scope="col">COD.</th>
+                            <th scope="col">COD. COMPRA</th>
+                            <th scope="col">NOME PROD.</th>
+                            <th scope="col">COD. PRODUTO</th>
                             <th scope="col">PREÇO UNID.(R$)</th>
                             <th scope="col">QTD. COMPRADA</th>
-                            <th scope="col">TOTAL(R$)</th>
-                            <th scope="col">DATA</th>
                             <th scope="col">FORNECEDOR</th>
+                            <th scope="col">TOTAL(R$)</th>
+                            <th scope="col">DATA</th>    
                         </tr>
                         </thead>
                         <tbody>
-                        <tr class="items-prods">
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td><a class="link-info" style="text-decoration:none;" href="/detalhes/">Baby Bliss</a></td>
-                            <td>Jurema</td>
-                            <td>2347</td>
-                            <td>43</td>
-                            <td>113</td>
-                            <td>200,00</td>
-                            <td>12/06/22</td>
-                            <td>Atacadão A</td>
-                        </tr>
-                        <tr>
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td><a class="link-info" style="text-decoration:none;" href="/detalhes/">Liquidificador</a></td>
-                            <td>Mondial</td>
-                            <td>2500</td>
-                            <td>50</td>
-                            <td>100</td>
-                            <td>200,00</td>
-                            <td>12/06/22</td>
-                            <td>Atacadão B</td>
-                        </tr>
-                        <tr>
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td ><a class="link-info" style="text-decoration:none;" href="/detalhes/">Faca Elétrica</a></td>
-                            <td>Wallita</td>
-                            <td>7659</td>
-                            <td>12</td>
-                            <td>24</td>
-                            <td>200,00</td>
-                            <td>12/06/22</td>
-                            <td>Atacadão C</td>
-                        </tr>
+                            {{#each compras}}
+                                <tr class="items-prods">
+                                    <th scope="row"> <input class="form-check-input" type="checkbox" value={{this.codCompra}} id="flexCheckDefault"></th>
+                                    <td><a class="link-info" style="text-decoration:none;" href="/detalhes/:codProduto" target="_self">{{this.codCompra}}</a></td>
+                                    <td>{{this.nome}}</td>
+                                    <td>{{this.codProduto}}</td>
+                                    <td>{{this.preco_unid_compra}}</td>
+                                    <td>{{this.quantidade}}</td>
+                                    <td>{{this.fornecedor}}</td>
+                                    <td>{{this.total}}</td>
+                                    <td>{{this.Data_hora}}</td>
+                                </tr>
+                            {{/each}}
                         </tbody>
                     </table>
             </div>
