+++ RCNutilidades/views/detalhes_itens.hbs
@@ -20,20 +20,12 @@
                         </tr>
                         </thead>
                         <tbody>
-                        <tr class="items-prods">
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td><a class="link-info" style="text-decoration:none;" href="">Baby Bliss</a></td>
-                           
-                        </tr>
-                        <tr>
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td><a class="link-info" style="text-decoration:none;" href="">Liquidificador</a></td>
-                            
-                        </tr>
-                        <tr>
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td ><a class="link-info" style="text-decoration:none;" href="">Faca Elétrica</a></td>
-                        </tr>
+                            {{#each nomeProds}}
+                                <tr class="items-prods">
+                                    <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
+                                    <td><a class="link-info" style="text-decoration:none;" href="/detalhes_dados/:{{this.codProduto}}">{{this.nome}}</a></td>
+                                </tr>
+                            {{/each}}
                         </tbody>
                     </table>
             </div>
