+++ RCNutilidades/views/index.hbs
@@ -9,7 +9,6 @@
     <link rel="stylesheet" href="/stylesheets/index.css"/>
 </head>
 <body class="body_index">
-
     <div class="header_geral">
         <div class="row">
             <div class="col-4">
@@ -56,9 +55,6 @@
                                 <a class="nav-link" aria-current="page" href="/estoque/" target="if-body">Estoque/Produtos</a>
                             </li>
                             <li class="nav-item">
-                                <a class="nav-link" href="/precos/" target="if-body">Preços</a>
-                            </li>
-                            <li class="nav-item">
                                 <a class="nav-link" href="/vendas/" target="if-body">Vendas</a>
                             </li>
                             <li class="nav-item">
@@ -69,10 +65,7 @@
             </div>    
         </div>
     </div>
-
    <iframe class="body-app" src="/estoque/" name="if-body" frameborder="0"></iframe>
-
-   {{!-- <iframe class="footer-app" src="/estoque/" name="if-body" frameborder="0"></iframe> --}}  
  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa" crossorigin="anonymous"></script>
 </body>
 </html>
