--- RCNutilidades/views/precos.hbs
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>RCN - Preços</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
-    <link rel="stylesheet" href="/stylesheets/index.css"/>
-</head>
-<body>
-    <div class="header-estoque">
-        <div class="row">
-            <h2 class="nome-pagina">Preços</h2>
-        </div>
-    
-        <div class="row">
-    
-            <div class="col-6">
-                <div class="dropdown">
-                    <button id="btn-organizar" class="btn btn-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
-                    Organizar
-                    </button>
-                    <ul class="dropdown-menu">
-                    <li><a class="dropdown-item" href="#">Mais recentes</a></li>
-                    <li><a class="dropdown-item" href="#">Mais antigos</a></li>
-                    <li><a class="dropdown-item" href="#">Maior valor</a></li>
-                    <li><a class="dropdown-item" href="#">Menor valor</a></li>
-                    </ul>
-                </div>
-            </div>
-    
-            <div class="col-6" id="btn-add"> 
-                <button type="button" class="btn btn-primary">+ Novo</button>
-                <button type="button" class="btn btn-danger"><img src="/images/lata-lixo.png" alt="lata-lixo" style="width: 20px; padding: 1px;"></button>
-            </div>
-        </div>
-    </div>
-     
-    <div class="lista-estoque">
-        <div class="row">
-            <div class="itens-prod">
-                    <table class="table table-hover">
-                        <thead>
-                        <tr class="header-prods">
-                            <th scope="col"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <th scope="col">NOME</th>
-                            <th scope="col">MARCA</th>
-                            <th scope="col">COD.</th>
-                            <th scope="col">PREÇO/COMPRA</th>
-                            <th scope="col">PREÇO/VENDA</th>
-                            <th scope="col">SETOR</th>
-                        </tr>
-                        </thead>
-                        <tbody>
-                        <tr class="items-prods">
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td><a class="link-info" style="text-decoration:none;" href="/detalhes/">Baby Bliss</a></td>
-                            <td>Jurema</td>
-                            <td>2347</td>
-                            <td>43</td>
-                            <td>113</td>
-                            <td>Eletro</td>
-                        </tr>
-                        <tr>
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td><a class="link-info" style="text-decoration:none;" href="/detalhes/">Liquidificador</a></td>
-                            <td>Mondial</td>
-                            <td>2500</td>
-                            <td>50</td>
-                            <td>100</td>
-                            <td>Eletro</td>
-                        </tr>
-                        <tr>
-                            <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
-                            <td ><a class="link-info" style="text-decoration:none;" href="/detalhes/">Faca Elétrica</a></td>
-                            <td>Wallita</td>
-                            <td>7659</td>
-                            <td>12</td>
-                            <td>24</td>
-                            <td>Eletro</td>
-                        </tr>
-                        </tbody>
-                    </table>
-            </div>
-        </div>
-    </div>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa" crossorigin="anonymous"></script>
-</body>
-</html>
