+++ front-end-rcn/tabelas_bk.html
@@ -1,0 +1,26 @@
+<!-- Estoque -->
+
+<tr>
+    <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
+    <td><a class="link-info" style="text-decoration:none;" href="/detalhes/" target="_self">Liquidificador</a></td>
+    <td>Mondial</td>
+    <td>2500</td>
+    <td>50</td>
+    <td>66</td>
+    <td>113</td>
+    <td>Eletro</td>
+    <td>unitário</td>
+    <td>Reposição - 12/06/22</td>
+</tr>
+<tr>
+    <th scope="row"> <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault"></th>
+    <td ><a class="link-info" style="text-decoration:none;" href="/detalhes/" target="_self">Faca Elétrica</a></td>
+    <td>Wallita</td>
+    <td>7659</td>
+    <td>12</td>
+    <td>66</td>
+    <td>113</td>
+    <td>Eletro</td>
+    <td>unitário</td>
+    <td>Venda - 12/06/22</td>
+</tr>
